moved fit functions to reconstruction
